+++ tests/DispatcherTest.php
@@ -10,7 +10,6 @@ use Firehed\API\Authorization;
 use Firehed\API\Interfaces\EndpointInterface;
 use Firehed\API\Errors\HandlerInterface;
 use Psr\Container\ContainerInterface;
-use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamInterface;
@@ -91,7 +90,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
     public function testSetRequestReturnsSelf()
     {
         $d = new Dispatcher();
-        $req = $this->createMock(RequestInterface::class);
+        $req = $this->createMock(ServerRequestInterface::class);
         $req->method('getHeaders')->willReturn([]);
         $req->method('getBody')->willReturn($this->createMock(StreamInterface::class));
         $req->method('getUri')->willReturn($this->createMock(UriInterface::class));
@@ -197,7 +196,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
     public function testPsr15()
     {
         $request = $this->createMock(ServerRequestInterface::class);
-        $modifiedRequest = $this->getMockRequestWithUriPath('/c', 'GET', [], ServerRequestInterface::class);
+        $modifiedRequest = $this->getMockRequestWithUriPath('/c', 'GET', []);
         $response = $this->createMock(ResponseInterface::class);
         $modifiedResponse = $this->createMock(ResponseInterface::class);
 
@@ -455,7 +454,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
             ->method('handleException')
             ->with($e)
             ->will($this->throwException($e));
-        $this->executeMockRequestOnEndpoint($endpoint, $dispatcher, ServerRequestInterface::class);
+        $this->executeMockRequestOnEndpoint($endpoint, $dispatcher);
     }
 
     /** @covers ::dispatch */
@@ -480,24 +479,6 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
         } catch (Throwable $t) {
             $this->assertSame($e, $t, 'A different exception was thrown');
         }
-    }
-
-    /** @covers ::setRequest */
-    public function testDeprecationWarningIsIssuedWithBaseRequest()
-    {
-        error_reporting($this->reporting); // Turn standard reporting back on
-        $dispatcher = new Dispatcher();
-        $this->expectException(\PHPUnit\Framework\Error\Deprecated::class);
-        $dispatcher->setRequest($this->createMock(RequestInterface::class));
-    }
-
-    /** @covers ::setRequest */
-    public function testDeprecationWarningIsNotIssuedWithServerRequest()
-    {
-        error_reporting($this->reporting); // Turn standard reporting back on
-        $dispatcher = new Dispatcher();
-        $dispatcher->setRequest($this->createMock(ServerRequestInterface::class));
-        $this->assertTrue(true, 'No error should have been raised');
     }
 
     /** @covers ::setAuthProviders */
@@ -530,7 +511,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
             ->method('authorize')
             ->willReturn(new Authorization\Ok());
 
-        $req = $this->getMockRequestWithUriPath('/c', 'GET', [], ServerRequestInterface::class);
+        $req = $this->getMockRequestWithUriPath('/c', 'GET', []);
         $routes = ['GET' => ['/c' => 'ClassThatDoesNotExist']];
         $endpoint = $this->createMock(Interfaces\AuthenticatedEndpointInterface::class);
 
@@ -575,7 +556,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
             HandlerInterface::class => $eh,
             'ClassThatDoesNotExist' => $ep,
         ]);
-        $req = $this->getMockRequestWithUriPath('/c', 'GET', [], ServerRequestInterface::class);
+        $req = $this->getMockRequestWithUriPath('/c', 'GET', []);
         $routes = ['GET' => ['/c' => 'ClassThatDoesNotExist']];
         $dispatcher = new Dispatcher();
         $dispatcher->setContainer($container)
@@ -610,7 +591,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
         $authz2->expects($this->never())
             ->method('authorize');
 
-        $req = $this->getMockRequestWithUriPath('/c', 'GET', [], ServerRequestInterface::class);
+        $req = $this->getMockRequestWithUriPath('/c', 'GET', []);
         $routes = ['GET' => ['/c' => 'ClassThatDoesNotExist']];
         $endpoint = $this->createMock(Interfaces\AuthenticatedEndpointInterface::class);
 
@@ -658,7 +639,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
 
         $dispatcher = new Dispatcher();
         $dispatcher->setAuthProviders($authn, $authz);
-        $res = $this->executeMockRequestOnEndpoint($endpoint, $dispatcher, ServerRequestInterface::class);
+        $res = $this->executeMockRequestOnEndpoint($endpoint, $dispatcher);
         $this->assertSame($response, $res);
     }
 
@@ -690,7 +671,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
         $dispatcher = new Dispatcher();
         $dispatcher->setAuthProviders($authn, $authz);
         try {
-            $this->executeMockRequestOnEndpoint($endpoint, $dispatcher, ServerRequestInterface::class);
+            $this->executeMockRequestOnEndpoint($endpoint, $dispatcher);
             $this->fail('An authorization exception should have been thrown');
         } catch (Authorization\Exception $e) {
             $this->assertSame($authzEx, $e);
@@ -721,7 +702,7 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
         $dispatcher = new Dispatcher();
         $dispatcher->setAuthProviders($authn, $authz);
         try {
-            $this->executeMockRequestOnEndpoint($endpoint, $dispatcher, ServerRequestInterface::class);
+            $this->executeMockRequestOnEndpoint($endpoint, $dispatcher);
             $this->fail('An exception should have been thrown');
         } catch (Authentication\Exception $e) {
             $this->assertSame($authnEx, $e);
@@ -751,48 +732,22 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
      * @param string $uri path component of URI
      * @param string $method optional HTTP method
      * @param array $query_data optional raw, unescaped query string data
-     * @param string $requestClass What RequestInterface to mock
-     * @return RequestInterface
+     * @return ServerRequestInterface
      */
     private function getMockRequestWithUriPath(
         string $uri,
         string $method = 'GET',
-        array $query_data = [],
-        string $requestClass = RequestInterface::class
-    ): RequestInterface {
-        if ($requestClass === RequestInterface::class) {
-            $request = new Request(
-                $uri.'?'.http_build_query($query_data),
-                $method
-            );
-        } elseif ($requestClass === ServerRequestInterface::class) {
-            $request = new ServerRequest(
-                [],
-                [],
-                $uri,
-                $method
-            );
-        } else {
-            throw new \Exception('Invalid request class');
-        }
+        array $query_data = []
+    ): ServerRequestInterface {
+        $uri .= '?' . http_build_query($query_data);
+        $request = new ServerRequest(
+            [],
+            [],
+            $uri,
+            $method,
+            'php://memory'
+        );
         return $request;
-        $mock_uri = $this->createMock(UriInterface::class);
-        $mock_uri->expects($this->any())
-            ->method('getPath')
-            ->will($this->returnValue($uri));
-        $mock_uri->method('getQuery')
-            ->will($this->returnValue(http_build_query($query_data)));
-
-        /** @var RequestInterface | \PHPUnit\Framework\MockObject\MockObject */
-        $req = $this->createMock($requestClass);
-
-        $req->expects($this->any())
-            ->method('getUri')
-            ->will($this->returnValue($mock_uri));
-
-        $req->method('getMethod')
-            ->will($this->returnValue($method));
-        return $req;
     }
 
     /**
@@ -822,10 +777,9 @@ class DispatcherTest extends \PHPUnit\Framework\TestCase
      */
     private function executeMockRequestOnEndpoint(
         EndpointInterface $endpoint,
-        Dispatcher $dispatcher = null,
-        string $requestClass = RequestInterface::class
+        Dispatcher $dispatcher = null
     ): ResponseInterface {
-        $req = $this->getMockRequestWithUriPath('/container', 'GET', [], $requestClass);
+        $req = $this->getMockRequestWithUriPath('/container', 'GET', []);
         $list = [
             'GET' => [
                 '/container' => 'ClassThatDoesNotExist',